--- frontend/src/metabase/writeback/api.ts
@@ -1,3 +1,0 @@
-import { POST } from "metabase/lib/api";
-
-export const deleteRow = POST("/api/actions/row/delete");
